Fix dependency error on express version
There is the following error when running the postman-local cli :

`TypeError: Missing parameter name at 1: https:///git.new/pathToRegexpError`. It is due to a new major version of `path-to-regexp` being used :
`path-to-regexp` is a dependency of `router`, itself being a dependency of `express`. Currently the `>=` in the version allows to get any version above the one specified.

By using `^` for the versions instead of `>=` we ensure that we are keeping the same major version.